drivers/base: cpu: Add node for core control isolation
This cpu node can be used to check the isolated cores
which are made unavailable for scheduler and exempted
from waking up on interrupts or from running timers at
the given time based on the load. All the runnable tasks,
timers and hrtimers should have migrated to other online
running cores.

Change-Id: I6771f53e59ca37b2cc3b17309cc8f4d75462cdfa
Signed-off-by: Sabyasachi Singh <sssingh@codeaurora.org>
Signed-off-by: Syed Rameez Mustafa <rameezmustafa@codeaurora.org>
Signed-off-by: Satya Durga Srinivasu Prabhala <satyap@codeaurora.org>