Revert "drm/msm/disp/dpu: fix early dereference of phys encoder"
This reverts commit 77a209cd7153cd758aa1b1417fb10017bead9454.

This reversion is happening as a result of disabling the
disp/dpu directory as it will not used in 4.19 downstream display
driver and these changes are made in the common layer which
affects the downstream driver functionalities. Separate patches
are submitted to revert the upstream driver and introduce the
downstream version of the display driver.

Change-Id: Idcdb854395fff490b96cda1b9b4eed375deb9b18
Signed-off-by: Samantha Tran <samtran@codeaurora.org>