Revert "drm/msm/disp/dpu: Mark a handful of functions as static"
This reverts commit fba33cae6aa3eaa3627a603d62d4652f12308b8d.

This reversion is happening as a result of disabling the
disp/dpu directory as it will not be used in 4.19 downstream
display driver and these changes are made in the common layer
which affects the downstream driver functionalities. Separate
patches are submitted to revert the upstream driver and
introduce the downstream version of the display driver.

Change-Id: I9703b55efa7074ae746b6418e27ef8ff84d58335
Signed-off-by: Samantha Tran <samtran@codeaurora.org>