Revert "drm/msm/disp/dpu: Remove unused funcs in dpu_formats.c"
This reverts commit f2c9a924c1acf42f692f2b5ac77531eea69460a3.

This reversion is happening as a result of disabling the
disp/dpu directory as it will not be used in 4.19 downstream
display driver and these changes are made in the common layer
which affects the downstream driver functionalities. Separate
patches are submitted to revert the upstream driver and
introduce the downstream version of the display driver.

Change-Id: I51dfc5a03ba127eb4aac343e8589a65fcb18dec9
Signed-off-by: Samantha Tran <samtran@codeaurora.org>