Revert "drm/msm: enable zpos normalization"
This reverts commit bb676df12b5e81cab57d1a212a6e9cfc343875a7.

This reversion is happening as a result of disabling the
disp/dpu directory as it will not be used in 4.19 downstream
display driver and these changes are made in a common layer
which affects the downstream driver functionalities. Separate
patches are submitted to revert the upstream driver and
introduce the downstream version of the display driver.

Change-Id: I96cf9c266c34ec83c545c3b486fe071b95d6fc0d
Signed-off-by: Samantha Tran <samtran@codeaurora.org>