power: qpnp-fg-gen4: Publish ESR SOH parameters only once during charging
Currently, ESR parameters for SOH (State of Health) algorithm
is published only for 50% SOC (State of Charge). However, if
the delta MSOC interrupt fires frequently and MSOC is rounded
close to 50%, then it can be published more than once during
a charging cycle. Allow it only once so that we can have valid
parameters for SOH calculation.

Change-Id: I5c3a63b4f4670250cc8af5ec5a0c438e31f2fbca
Signed-off-by: Subbaraman Narayanamurthy <subbaram@codeaurora.org>