power: smb5-lib: Fix no charging with Type-C legacy cable via PD adaptor
A no charging issue is observed and can be reproduced by attaching the
phone with Type-C/PD adaptor first, and then inserting a Type-C legacy
charger cable to it.

The zero charging current is due to USB remaining in suspend state after
TypeC mode is switched from source to sink mode. This is currently not
handled when ICL is set with USB5/1(100mA/500mA) or USB9(150mA). In this
case, PD votes 500mA for USB ICL.

Fix it.

CRs-Fixed: 2360281
Change-Id: I438dfed6075e26d83edb4bceeaaae43e8ac96b6e
Signed-off-by: Harry Yang <harryy@codeaurora.org>