drm: Read dp mst edid immediately after plugin
In the current implementation, if a new dp
connector is plugged in to an existing port, the
edid is not read during plugin time. The edid is
read when userspace calls get_connector(), while
kernel is holding the atomic modeset lock. Since
the get_edid call may take a couple of seconds,
no frames can be committed during this time,
which causes a momentary freeze.

Fix this issue by reading and caching the edid as
soon as the connector is plugged in. This will
prevent the atomic commit from blocking,
resulting in no freeze upon plugin.

CRs-Fixed: 2496461
Change-Id: I9414ce072c8e6e4919c15291331e9e08b9030460
Signed-off-by: Fuad Hossain <fhossain@codeaurora.org>