mmc: core: dont send PON during suspend
During suspend, eMMC VCC can be turned off. So instead of sending
a power off notification, send a sleep command. According to the
eMMC 4.5 specification, this is the only time where its clearly
mentioned that the VCC regulator can be turned OFF.

This has been clarified in eMMC 5.0 specification and this applies
even if PON with sleep is not sent to the card.

Change-Id: Id30ca76e8953ddb7a77e8fcfd4ec01469ceeb068
Signed-off-by: Krishna Konda <kkonda@codeaurora.org>
Signed-off-by: Maya Erez <merez@codeaurora.org>
Signed-off-by: Ram Prakash Gupta <rampraka@codeaurora.org>