mmc: cqhci: Increase HAC interrupt wait time in halt path
When system is heavily loaded, in some cases interrupt servicing
is getting effected and cmdq halt interrupt handler is getting invoked
after 1 sec delay. Since wait time of HAC interrupt in cmdq driver
is 1 sec, the delayed interrupt is being treated as unexpected
interrupt.
For fixing this case, increasing the timeout to 10 seconds.

Change-Id: I3f6aac1779ad5b84bc8bdf24dad1c72a2d8f5957
Signed-off-by: Veerabhadrarao Badiganti <vbadigan@codeaurora.org>
Signed-off-by: Ram Prakash Gupta <rampraka@codeaurora.org>