power: smb5: Handle WLS input during initial status
Currently, only the DC plugin interrupt triggers the configuration of
WLS input and the running of DCIN AICL. Allow these operations to occur
as part of the initial set of operations the driver does as soon as it
comes up to handle the case when the SoC is powered on by WLS input.

Also check for DC PSY's presence before calling power_supply_changed()
on it.

Change-Id: If4b9847ba32f7e35eb7a4424635209aeeaef7453
Signed-off-by: Guru Das Srinagesh <gurus@codeaurora.org>