mink: handle error in local object response
When HLOS receives inbound request, it processes req and sends a
response. If for some reason, trusted execution environment finds
an error in it, it sends inbound request again with error set in
ret[1]. However, HLOS does not look at ret[1] because it is for
userspace's consumption. This results in an infinite loop.
As a fix, HLOS would look into ret[1] if it is a local object.

Change-Id: I189cad47ee95f6ea9a2ddb5507ad0ee7de9c60ec
Signed-off-by: Dinesh K Garg <dineshg@codeaurora.org>