usb: phy: qusb: Prevent charger detection if EUD is enabled
With EUD enabled in boot, the charger detection would
cause failure of EUD enumeration on PC because it
involves resetting the PHY and disabling the clamp
register.

So, prevent charger detection if EUD is enabled, and
just notify the charger type as SDP to the PMIC.
While at it, prevent the clearing of clamp register
which might also cause the EUD to fail on PC.

While at it, refactor the code for setting and clearing
the TCSR clamp register to prevent multiple "if"
conditions repeating over the driver.

Change-Id: Ia8c57ea2c0e01ae337c53107d2d7e60b873df20c
Signed-off-by: Sriharsha Allenki <sallenki@codeaurora.org>